Added a game loop
So, now, a game loop.  Touch all candles to light them, which lights
the master candle and makes the player blinder.  After 10 levels of
blindness, the game ends.  What happens at the end is not defined,
other than a wonGame method in CandleManager.  CandleManager is
basically the master game loop.